+++ CHANGES.rst
@@ -5,6 +5,8 @@
 
 - Added new `mock_observables` functions `radial_distance` and `radial_distance_and_velocity` functions. See https://github.com/astropy/halotools/pull/782
 
+- Added new `distribution_matching_indices` function. See https://github.com/astropy/halotools/pull/795.
+
 - Removed `max_sample_size` keyword argument from all `mock_observables` functions.